detect all 3-fold-repetitions in whole match do not remember complete evaluated board positions but only repeatable positions do not use mutable static value vectors for min and max

+++ src/main/java/de/thohee/useless/chess/board/FENParser.java
@@ -103,7 +103,7 @@ public class FENParser {
 
 	private static void setDepth(String field, BoardPosition boardPosition) {
 		int nextFullMove = Integer.parseInt(field);
-		long depth = 2 * (nextFullMove - 1);
+		int depth = 2 * (nextFullMove - 1);
 		if (boardPosition.getColourToMove().equals(Colour.Black)) {
 			++depth;
 		}
